requirements.txt
numpy
scipy
scikit-learn
statsmodels
requests
httpx
fastapi
uvicorn
anthropic
openai
python-dotenv
